component: Implement product list component
The component get a list of all the product from the db and renders them if exists, when clicking on the delete button the product id is being passed to a func who handles the delete and deletes the product. when the product div is being click the product the specific product is being fetch and set to setProductById props that will pass to a different component.

Updated(Temp): Page top header design