feat: Products Pagination
Now when products are being fetched we also set a pagination info state.

created a product pagination component, the change and switch page function were created in productMain for passing the page number as props.

change all fetchAllProducts functions inside update, add and delete to switchToPage so when usings any of these functions the user wont be force to go to the first page.