component: Implement pagination
Now when products are being fetched we also set a pagination info state.

created a product pagination component, the change and switch page function were created in parent product component for passing the page number as props.

change all fetchAllProducts functions inside update, add and delete to switchToPage so when using any of these functions the user won't be forced to go to the first page.